next: sdk: remove global rooms and helpers
Now that presence is fetched from PFS, we don't need to join discovery
global room anymore, and since PFS and MS rooms were removed previously,
we can clear the utility and helpers which supported this.
